requirements.txt
appdirs==1.4.4
arrow==1.0.0
attrs==20.3.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
feedgen==0.9.0
idna==2.10
lxml==4.6.2
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
soupsieve==2.2
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
